Reorganize beamline elements
